requirements.txt
langchain
langchain-ollama
langchain-chroma
panda
